Added MIDI learn functionality and reset.
If the main switch is pressed at power on, default MIDI channel 1 and controller 20 are stored.
If the main switch is pressed during operation and the input jack is not connected, MIDI learn is started. In this case, the device waits till it receives the first CC message. The channel and controller are stored afterwards.