Addition of periodic UART overrun check.
UART gets into overrun condition once it receives data and the previous data has not been consumed by the application yet. In this case the ORE bit (bit 3) in the ISR register is set to 1 which disables all further interrupts. As this can happen at any time, we periodically check the ORE bit and re-issue the interrupt in case it is set to high.
This can broken interrupts after midi learn and write to flash. During write to flash, the interrupts and therefore the data are not consumed which leads to broken interupts.